Refactor: consolidate e2e helpers; use IsBlocked in SpawnVillagerAtHouse
- Extract all shared e2e_tests helper functions (tick, drainOffers,
  tickDraining, moveDir, moveSafe, sendKey, statusBar, viewLines,
  charAtScreen, stripANSI) into a new helpers_test.go, removing the
  duplication across log_storage_test.go and house_test.go.
- Replace manual tile nil/structure check in SpawnVillagerAtHouse with
  World.IsBlocked, which is the established API for this exact predicate.
